requirements.txt
ccxt>=4.0.0
pandas>=1.5.0
numpy>=1.24.0
talib>=0.4.25
requests>=2.28.0
python-dotenv>=1.0.0